requirements.txt
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.26.0
psutil==5.9.4
py-machineid==0.3.0
pymongo==4.3.3
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.2
winregistry==1.1.1
WTForms==3.0.1
zipp==3.13.0
